wip android

+++ projects/mouse-android-emulator/event-dispatcher/android/app/src/main/java/com/example/remotemouse/AvailableConnectionData.java
@@ -5,7 +5,7 @@ public class AvailableConnectionData {
     public final String hostname;
     public final int port;
 
-    public AvailableConnectionData(String ip, String hostname, int port) {
+    public AvailableConnectionData(String ip, String hostname, short port) {
         this.ip = ip;
         this.hostname = hostname;
         this.port = port;
